+++ scripts/react-app-ssr-vite/server.js
@@ -7,9 +7,6 @@ import { getPageData, renderViteHtml } from '@kibalabs/build/scripts/react-app-s
 import compression from 'compression';
 import express from 'express';
 
-import * as app from './_ssr/assets/App.js';
-import data from './data.json' with { type: 'json' };
-
 const shouldCompress = (req, res) => {
   if (req.headers['x-no-compression']) {
     return false;
@@ -18,17 +15,25 @@ const shouldCompress = (req, res) => {
 };
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const htmlTemplate = await fs.readFileSync(path.join(__dirname, '_client/index.html'), 'utf-8');
+const clientDirectory = path.join(__dirname, '_client');
+const ssrDirectory = path.join(__dirname, '_ssr');
+const app = (await import(path.join(ssrDirectory, 'assets/app.js')));
+const appData = (await import(path.join(ssrDirectory, 'appData.json'), { with: { type: 'json' } })).default;
+const htmlTemplate = await fs.readFileSync(path.join(clientDirectory, 'index.html'), 'utf-8');
 export const createAppServer = () => {
   const server = express();
   server.disable('x-powered-by');
-  server.use(express.static(path.join(__dirname, '_client'), { immutable: true, maxAge: '1y', index: false }));
+  server.use(express.static(clientDirectory, { immutable: true, maxAge: '1y', index: false }));
   server.use(compression({ filter: shouldCompress }));
   server.get('*', async (req, res) => {
+    const queryString = Object.entries(req.query)
+      .map(([key, value]) => `${encodeURIComponent(key)}=${encodeURIComponent(value)}`)
+      .join('&');
+    console.log(`${req.method}:${req.path}:${queryString}`);
     const startTime = new Date();
     const page = { path: req.path };
-    const pageData = await getPageData(req.path, app.routes, app.globals);
-    const html = renderViteHtml(app.App, page, data.defaultSeoTags, data.name, pageData, htmlTemplate);
+    const pageData = (app.routes && app.globals) ? await getPageData(page.path, app.routes, app.globals) : null;
+    const html = await renderViteHtml(app.App, page, appData.defaultSeoTags, appData.name, pageData, htmlTemplate);
     // TODO(krishan711): move this stuff to a middleware
     if (process.env.NAME) {
       res.header('X-Server-Name', process.env.NAME);
@@ -41,14 +46,14 @@ export const createAppServer = () => {
     }
     const duration = (new Date() - startTime) / 1000.0;
     res.header('X-Response-Time', String(duration));
-    console.log(req.method, req.path, req.query, res.statusCode, duration);
+    console.log(`${req.method}:${req.path}:${queryString}:${res.statusCode}:${duration}`);
     res.status(200).set({ 'Content-Type': 'text/html' }).end(html);
   });
   return server;
 };
 
 const host = '0.0.0.0';
-const port = data.port || 3000;
+const port = appData.port || 3000;
 const server = createAppServer();
 server.listen(port, host, () => {
   console.log(`Started server at http://${host}:${port}`);