+++ scripts/react-app-static-vite/build.js
@@ -6,19 +6,18 @@ import { build, mergeConfig } from 'vite';
 
 import { getPageData, renderViteHtml } from '../react-app-static/static.js';
 import { buildReactAppViteConfig } from '../react-app-vite/app.config.js';
-import { removeUndefinedProperties, runParamsConfigModifier } from '../util.js';
+import { buildParams } from '../util.js';
 
 
 export const buildStaticReactApp = async (inputParams = {}) => {
   const defaultParams = {
-    dev: false,
     configModifier: undefined,
     polyfill: true,
     polyfillTargets: undefined,
     webpackConfigModifier: undefined,
     analyzeBundle: false,
     shouldAliasModules: true,
-    addHtmlOutput: false,
+    addHtmlOutput: true,
     addRuntimeConfig: true,
     runtimeConfigVars: {},
     seoTags: [],
@@ -28,8 +27,7 @@ export const buildStaticReactApp = async (inputParams = {}) => {
     publicDirectory: path.join(process.cwd(), './public'),
     appEntryFilePath: path.join(process.cwd(), './src/App.tsx'),
   };
-  let params = { ...defaultParams, ...removeUndefinedProperties(inputParams) };
-  params = await runParamsConfigModifier(params);
+  const params = await buildParams(defaultParams, inputParams, false);
   let viteConfig = buildReactAppViteConfig(params);
   if (params.viteConfigModifier) {
     viteConfig = params.viteConfigModifier(viteConfig);
@@ -39,19 +37,19 @@ export const buildStaticReactApp = async (inputParams = {}) => {
 
   const outputDirectoryPath = path.resolve(params.outputDirectory);
   const appEntryFilePath = path.resolve(params.appEntryFilePath);
-  const clientOutputDirectoryPath = path.join(outputDirectoryPath, '_client');
-  const ssrOutputDirectoryPath = path.join(outputDirectoryPath, '_ssr');
+  const clientDirectory = path.join(outputDirectoryPath, '_client');
+  const ssrDirectory = path.join(outputDirectoryPath, '_ssr');
   console.log('building app...');
   await build(mergeConfig(viteConfig, {
     build: {
-      outDir: clientOutputDirectoryPath,
+      outDir: clientDirectory,
     },
   }));
   console.log('building server app...');
   await build(mergeConfig(viteConfig, {
     build: {
       ssr: true,
-      outDir: ssrOutputDirectoryPath,
+      outDir: ssrDirectory,
       rollupOptions: {
         input: appEntryFilePath,
         // NOTE(krishan711): prevent the hashes in the names
@@ -63,17 +61,17 @@ export const buildStaticReactApp = async (inputParams = {}) => {
       },
     },
   }));
-
-  const app = (await import(path.join(ssrOutputDirectoryPath, 'assets/App.js')));
-  const htmlTemplate = await fs.readFileSync(path.join(clientOutputDirectoryPath, 'index.html'), 'utf-8');
+  const app = (await import(path.join(ssrDirectory, 'assets/app.js')));
+  const appData = { name, port: params.port, defaultSeoTags: params.seoTags };
+  const htmlTemplate = await fs.readFileSync(path.join(clientDirectory, 'index.html'), 'utf-8');
   // NOTE(krishan711): if this could be done in an parallel way it would be faster!
   params.pages.forEach(async (page) => {
     console.log(`Rendering page ${page.path} to ${page.filename}`);
     const pageData = (app.routes && app.globals) ? await getPageData(page.path, app.routes, app.globals) : null;
-    const output = renderViteHtml(app.App, page, params.seoTags, name, pageData, htmlTemplate);
+    const html = await renderViteHtml(app.App, page, appData.defaultSeoTags, appData.name, pageData, htmlTemplate);
     const outputPath = path.join(outputDirectoryPath, page.filename);
     fs.mkdirSync(path.dirname(outputPath), { recursive: true });
-    fs.writeFileSync(outputPath, output);
+    fs.writeFileSync(outputPath, html);
     console.log(`Done rendering page ${page.path}`);
   });
 };