+++ scripts/react-app-static/static.js
@@ -4,6 +4,8 @@ import React from 'react';
 import { ChunkExtractor, ChunkExtractorManager } from '@loadable/server';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import ReactDOMServer from 'react-dom/server';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import ReactDOMStatic from 'react-dom/static';
 import { matchPath } from 'react-router';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { ServerStyleSheet, StyleSheetManager } from 'styled-components';
@@ -27,54 +29,104 @@ export const getPageData = async (urlPath, appRoutes, globals) => {
   return null;
 };
 
-export const renderViteHtml = (app, page, defaultSeoTags, appName, pageData, htmlTemplate) => {
-  let pageHead = { headId: '', base: null, title: null, links: [], metas: [], styles: [], scripts: [], noscripts: [] };
-  const setHead = (newHead) => { pageHead = newHead; };
+async function renderToString(reactComponent) {
+  const { prelude } = await ReactDOMStatic.prerenderToNodeStream(reactComponent);
+  return new Promise((resolve, reject) => {
+    let data = '';
+    prelude.on('data', (chunk) => {
+      data += chunk;
+    });
+    prelude.on('end', () => resolve(data));
+    prelude.on('error', reject);
+  });
+}
+
+
+const extractTitleTagFromString = (htmlString) => {
+  const titleRegex = /<title>.*?<\/title>/gi;
+  let lastTitleMatch = null;
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    const titleMatch = titleRegex.exec(htmlString);
+    if (titleMatch === null) {
+      break;
+    }
+    lastTitleMatch = titleMatch;
+  }
+  return lastTitleMatch ? lastTitleMatch[0] : null;
+};
+
+const extractLinkTagsFromString = (htmlString) => {
+  const linkTags = [];
+  const linkRegex = /<link.*?\/>/gi;
+  let linkMatch;
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    linkMatch = linkRegex.exec(htmlString);
+    if (linkMatch === null) {
+      break;
+    }
+    linkTags.push(linkMatch[0]);
+  }
+  return linkTags;
+};
+
+const extractHeadContentFromString = (htmlString) => {
+  const titleTag = extractTitleTagFromString(htmlString);
+  const linkTags = extractLinkTagsFromString(htmlString);
+  return linkTags.concat(titleTag ? [titleTag] : []);
+};
+
+const removeTitleTagFromString = (htmlString) => {
+  const titleTag = extractTitleTagFromString(htmlString);
+  if (titleTag == null) {
+    return htmlString;
+  }
+  return htmlString.replace(titleTag, '');
+};
+
+const removeStringsFromString = (string, stringsToRemove) => {
+  let output = string;
+  stringsToRemove.forEach((stringToRemove) => {
+    output = output.replace(stringToRemove, '');
+  });
+  return output;
+};
+
+export const renderViteHtml = async (app, page, defaultSeoTags, appName, pageData, htmlTemplate) => {
   const styledComponentsSheet = new ServerStyleSheet();
-  const bodyString = ReactDOMServer.renderToString(
+  const bodyString = await renderToString(
     React.createElement(
       StyleSheetManager,
       { sheet: styledComponentsSheet.instance },
-      React.createElement(app, { staticPath: page.path, pageData, setHead }),
+      React.createElement(app, { staticPath: page.path, pageData }),
     ),
   );
+  // NOTE(krishan711): prerenderToNodeStream doesnt extract out the stuff that should go to the head so we have to do it manually for now
+  const extractedHeadTags = extractHeadContentFromString(bodyString);
+  const extractedHeadTagsString = extractedHeadTags.join('');
+  const cleanedBodyString = removeStringsFromString(bodyString, extractedHeadTags);
   let pageSeoTags = page.seoTags;
   if (!pageSeoTags && defaultSeoTags && page.path === '/') {
     pageSeoTags = defaultSeoTags;
   }
   const seoTags = pageSeoTags || [];
-  const tags = [
-    ...(pageHead.title ? [pageHead.title] : [{ type: 'title', content: page.title || appName, attributes: [] }]),
-    ...(pageHead.base ? [pageHead.base] : []),
-    ...pageHead.links,
-    ...pageHead.metas,
-    ...pageHead.styles,
-    ...pageHead.scripts,
-  ];
   const headString = ReactDOMServer.renderToStaticMarkup(
     React.createElement(
-      'head',
+      React.Fragment,
       null,
       // NOTE(krishan711): this should be kept in sync with react-app/index.html
       React.createElement('script', { type: 'text/javascript', dangerouslySetInnerHTML: { __html: `window.KIBA_RENDERED_PATH = '${page.path}';` } }),
       React.createElement('script', { type: 'text/javascript', dangerouslySetInnerHTML: { __html: `window.KIBA_PAGE_DATA = ${JSON.stringify(pageData)};` } }),
       ...seoTags.map((tag) => React.createElement(tag.tagName, { ...tag.attributes, key: JSON.stringify(tag) })),
-      ...tags.map((tag) => React.createElement(tag.type, { ...tag.attributes, key: tag.headId, 'ui-react-head': tag.headId }, tag.content)),
-      // ...extractor.getLinkElements(),
-      // ...extractor.getStyleElements(),
       styledComponentsSheet.getStyleElement(),
     ),
   );
-  // const bodyAssetsString = ReactDOMServer.renderToStaticMarkup(
-  //   React.createElement(
-  //     React.Fragment,
-  //     null,
-  //     ...extractor.getScriptElements(),
-  //   ),
-  // );
-  let output = htmlTemplate.replace('<!--ssr-body-->', bodyString);
-  output = output.replace('<!--ssr-head-->', headString);
-  // output = output.replace(`<!--ssr-body-assets-->`, bodyAssetsString);
+  const cleanedHeadString = extractedHeadTagsString.includes('<title>') ? removeTitleTagFromString(headString) : headString;
+  const fullHeadString = extractedHeadTagsString + cleanedHeadString;
+  const cleanedHTmlTemplate = fullHeadString.includes('<title>') ? removeTitleTagFromString(htmlTemplate) : htmlTemplate;
+  let output = cleanedHTmlTemplate.replace('<!--ssr-body-->', cleanedBodyString);
+  output = output.replace('<!--ssr-head-->', fullHeadString);
   return output;
 };