+++ scripts/react-app-vite/createRuntimeConfigPlugin.js
@@ -1,7 +1,7 @@
 export const createRuntimeConfigPlugin = ({
   vars = {},
 }) => {
-  let scriptContent = 'const GLOBAL = typeof window !== "undefined" ? window : global;';
+  let scriptContent = 'const GLOBAL = typeof window !== "undefined" ? window : global;\n';
   Object.keys(vars).forEach((key) => {
     scriptContent += `GLOBAL.${key} = ${JSON.stringify(vars[key])};\n`;
   });