+++ scripts/react-app-vite/run.js
@@ -5,7 +5,6 @@ import { buildReactApp } from './build.js';
 
 const params = program
   .option('-c, --config-modifier <path>')
-  .option('-d, --dev')
   .option('-s, --start')
   .parse(process.argv)
   .opts();